refactor utility method of sorting view into new utility class
