+++ app/src/main/java/com/example/brennan/lister/ListActivity.java
@@ -47,6 +47,7 @@ import java.util.Calendar;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
+import com.example.brennan.lister.util.Utility;
 
 public class ListActivity extends AppCompatActivity {
     final Context context = this;
@@ -203,10 +204,10 @@ public class ListActivity extends AppCompatActivity {
             openCalendarActivity(findViewById(id));
         }
         else if (id == R.id.action_sort_desc){
-            sortByPriorityDesc();
+            Utility.sortByPriorityDesc(taskAdapter);
         }
         else if (id == R.id.action_sort_asc){
-            sortByPriorityAsc();
+            Utility.sortByPriorityAsc(taskAdapter);
         }
         return super.onOptionsItemSelected(item);
     }
@@ -318,27 +319,6 @@ public class ListActivity extends AppCompatActivity {
             allValid = false;
         }
         return allValid;
-    }
-
-
-    public void sortByPriorityDesc(){
-        taskAdapter.sort(new Comparator<Task>() {
-            @Override
-            public int compare(Task t1, Task t2) {
-                return t2.getPriorityInt() - (t1.getPriorityInt());
-            }
-        });
-        taskAdapter.notifyDataSetChanged();
-    }
-
-    public void sortByPriorityAsc(){
-        taskAdapter.sort(new Comparator<Task>() {
-            @Override
-            public int compare(Task t1, Task t2) {
-                return t1.getPriorityInt() - (t2.getPriorityInt());
-            }
-        });
-        taskAdapter.notifyDataSetChanged();
     }
 
     /**----------------------------------------------
